feat: rework how accounts are linked and managed
Adding an account now starts from a chooser that offers linking a bank
or entering the details by hand, so the two paths are equally visible
instead of linking hiding behind a second button.

The bank handshake runs full-screen with no app chrome behind it, and
the matching step that follows returns to the normal layout, with each
account shown as its number, name, type and balance before you decide
whether to create it or attach it to an account you already track.

Linked accounts are read-only apart from their name and your own notes,
since the institution owns the rest. Connections are managed from
Settings, where unlinking names every account it will disconnect, and a
connection that needs renewing is called out there and in the accounts
list.
