+++ messages/en.json
@@ -85,32 +85,31 @@
 	"accounts_status_excluded_description": "Excluded accounts are omitted from totals",
 	"accounts_status_closed_label": "Closed",
 	"accounts_status_closed_description": "Closed accounts are archived and excluded from totals",
+	"accounts_status_linked_label": "Linked",
+	"accounts_status_reauth_label": "Reconnection required",
 	"accounts_balance_tooltip_closed": "This account is closed and excluded from totals",
 	"accounts_balance_tooltip_excluded": "This account is excluded from totals",
 	"accounts_add_success": "Account added",
 	"accounts_add_failed": "Couldn't add account",
 	"accounts_add_page_title": "Add account",
+	"accounts_add_manual_page_title": "Add manually",
+	"accounts_add_choice_link_description": "Sign in to your bank so balances and transactions stay up to date on their own",
+	"accounts_add_choice_manual_description": "Enter the account details yourself and update its balance whenever you want",
+	"accounts_linked_institutions_link": "Linked institutions",
 	"accounts_link_page_title": "Link account",
-	"accounts_link_connect_section_title": "Connect with Plaid",
-	"accounts_link_setup_section_title": "Plaid setup",
 	"accounts_link_match_section_title": "Match accounts",
-	"accounts_link_not_configured": "Plaid is not configured on this server",
-	"accounts_link_configuration": "Set {clientId} and {secret}; {environment} is optional",
-	"accounts_link_documentation": "Plaid documentation",
+	"accounts_link_unavailable": "Account linking is not set up on this server",
 	"accounts_link_start_failed": "Couldn't start Plaid Link",
 	"accounts_link_exchange_failed": "Couldn't connect the Plaid account",
 	"accounts_link_save_failed": "Couldn't save the linked accounts",
 	"accounts_link_sync_failed": "Couldn't sync the linked accounts",
 	"accounts_link_success": "Accounts linked",
-	"accounts_link_no_mask": "No account number",
 	"accounts_link_label_account": "Account",
 	"accounts_link_label_type": "Type",
-	"accounts_link_label_account_number": "Account number",
 	"accounts_link_label_link_to": "Link to",
 	"accounts_link_create_new": "Create new account",
 	"accounts_link_existing_duplicate": "Each Plaid account must link to a different existing account",
 	"accounts_link_empty": "Plaid didn't return any accounts",
-	"accounts_link_back": "Back to accounts",
 	"accounts_link_cancel": "Cancel",
 	"accounts_link_confirm": "Confirm",
 	"accounts_link_confirming": "Linking accounts",
@@ -120,9 +119,6 @@
 	"accounts_link_discard_confirm_continue": "Discard connection",
 	"accounts_connection_section_title": "Connection",
 	"accounts_connection_institution_label": "Institution",
-	"accounts_connection_status_label": "Status",
-	"accounts_connection_status_error": "Last sync had errors",
-	"accounts_connection_status_reauth_required": "Reconnection required",
 	"accounts_connection_last_synced_label": "Last synced",
 	"accounts_connection_next_sync_label": "Next scheduled sync",
 	"accounts_connection_next_sync_paused": "Paused until you reconnect this bank",
@@ -144,12 +140,6 @@
 	"accounts_connection_unlink_success_one": "Connection removed from 1 account",
 	"accounts_connection_unlink_success_other": "Connection removed from {count} accounts",
 	"accounts_connection_unlink_failed": "Couldn't unlink the connection",
-	"accounts_sync_button": "Sync",
-	"accounts_sync_pending": "Syncing connections",
-	"accounts_sync_success": "Sync complete: {created} created, {skipped} skipped",
-	"accounts_sync_partial": "Sync finished with errors: {created} created, {skipped} skipped, {failed} failed",
-	"accounts_sync_in_progress": "A sync is already in progress",
-	"accounts_sync_failed": "Couldn't sync your connections",
 	"accounts_edit_success": "Account updated",
 	"accounts_edit_failed": "Couldn't update account",
 	"accounts_balance_updated": "Balance updated",
@@ -189,6 +179,7 @@
 	"accounts_overview_view_all_positions_other": "View all {count} positions",
 	"accounts_label_name": "Name",
 	"accounts_label_institution": "Institution",
+	"accounts_label_account_number": "Account number",
 	"accounts_label_category": "Category",
 	"accounts_label_balance_group": "Balance group",
 	"accounts_label_currency": "Currency",
@@ -648,12 +639,13 @@
 	"settings_agent_access_section_title": "AI agents",
 	"settings_agent_access_url_label": "Instructions URL",
 	"settings_connections_section_title": "Connections",
+	"settings_connections_nav_title": "Linked institutions",
 	"settings_connections_empty": "No banks connected yet",
 	"settings_connections_unnamed_institution": "Unnamed institution",
-	"settings_connections_schedule": "Syncs daily at {time}",
 	"settings_connections_table_header_institution": "Institution",
 	"settings_connections_table_header_status": "Status",
 	"settings_connections_table_header_last_synced": "Last synced",
+	"settings_connections_table_header_next_sync": "Next auto-sync",
 	"settings_connections_table_header_accounts": "Accounts",
 	"settings_connections_status_ok": "Connected",
 	"settings_connections_status_error": "Last sync had errors",
