+++ messages/es.json
@@ -85,31 +85,26 @@
 	"accounts_status_excluded_description": "Las cuentas excluidas no se consideran en los totales",
 	"accounts_status_closed_label": "Cerrada",
 	"accounts_status_closed_description": "Las cuentas cerradas están archivadas y se excluyen de los totales",
+	"accounts_status_linked_label": "Vinculada",
+	"accounts_status_reauth_label": "Requiere volver a conectarse",
 	"accounts_balance_tooltip_closed": "Esta cuenta está cerrada y se excluye de los totales",
 	"accounts_balance_tooltip_excluded": "Esta cuenta se excluye de los totales",
 	"accounts_add_success": "Cuenta agregada",
 	"accounts_add_failed": "No se pudo agregar la cuenta",
 	"accounts_link_page_title": "Vincular cuenta",
-	"accounts_link_connect_section_title": "Conectar con Plaid",
-	"accounts_link_setup_section_title": "Configuración de Plaid",
 	"accounts_link_match_section_title": "Emparejar cuentas",
-	"accounts_link_not_configured": "Plaid no está configurado en este servidor",
-	"accounts_link_configuration": "Configura {clientId} y {secret}; {environment} es opcional",
-	"accounts_link_documentation": "Documentación de Plaid",
+	"accounts_link_unavailable": "La vinculación de cuentas no está configurada en este servidor",
 	"accounts_link_start_failed": "No se pudo iniciar Plaid Link",
 	"accounts_link_exchange_failed": "No se pudo conectar la cuenta de Plaid",
 	"accounts_link_save_failed": "No se pudieron guardar las cuentas vinculadas",
 	"accounts_link_sync_failed": "No se pudieron sincronizar las cuentas vinculadas",
 	"accounts_link_success": "Cuentas vinculadas",
-	"accounts_link_no_mask": "Sin número de cuenta",
 	"accounts_link_label_account": "Cuenta",
 	"accounts_link_label_type": "Tipo",
-	"accounts_link_label_account_number": "Número de cuenta",
 	"accounts_link_label_link_to": "Vincular a",
 	"accounts_link_create_new": "Crear cuenta nueva",
 	"accounts_link_existing_duplicate": "Cada cuenta de Plaid debe vincularse a una cuenta existente diferente",
 	"accounts_link_empty": "Plaid no devolvió ninguna cuenta",
-	"accounts_link_back": "Volver a cuentas",
 	"accounts_link_cancel": "Cancelar",
 	"accounts_link_confirm": "Confirmar",
 	"accounts_link_confirming": "Vinculando cuentas",
@@ -119,9 +114,6 @@
 	"accounts_link_discard_confirm_continue": "Descartar conexión",
 	"accounts_connection_section_title": "Conexión",
 	"accounts_connection_institution_label": "Institución",
-	"accounts_connection_status_label": "Estado",
-	"accounts_connection_status_error": "La última sincronización tuvo errores",
-	"accounts_connection_status_reauth_required": "Requiere volver a conectarse",
 	"accounts_connection_last_synced_label": "Última sincronización",
 	"accounts_connection_next_sync_label": "Próxima sincronización programada",
 	"accounts_connection_next_sync_paused": "En pausa hasta que vuelvas a conectar este banco",
@@ -143,12 +135,6 @@
 	"accounts_connection_unlink_success_one": "Conexión eliminada de 1 cuenta",
 	"accounts_connection_unlink_success_other": "Conexión eliminada de {count} cuentas",
 	"accounts_connection_unlink_failed": "No se pudo desvincular la conexión",
-	"accounts_sync_button": "Sincronizar",
-	"accounts_sync_pending": "Sincronizando conexiones",
-	"accounts_sync_success": "Sincronización completada: {created} creados, {skipped} omitidos",
-	"accounts_sync_partial": "Sincronización terminada con errores: {created} creados, {skipped} omitidos, {failed} fallidos",
-	"accounts_sync_in_progress": "Ya hay una sincronización en curso",
-	"accounts_sync_failed": "No se pudieron sincronizar las conexiones",
 	"accounts_balance_updated": "Saldo actualizado",
 	"accounts_balance_failed": "No se pudo actualizar el saldo",
 	"accounts_share_created": "Compartido creado",
@@ -167,6 +153,10 @@
 	"accounts_share_leave_confirm_cancel": "Cancelar",
 	"accounts_share_leave_confirm_continue": "Continuar",
 	"accounts_add_page_title": "Agregar cuenta",
+	"accounts_add_manual_page_title": "Agregar manualmente",
+	"accounts_add_choice_link_description": "Inicia sesión en tu banco para que los saldos y las transacciones se actualicen solos",
+	"accounts_add_choice_manual_description": "Ingresa los datos de la cuenta y actualiza su saldo cuando quieras",
+	"accounts_linked_institutions_link": "Instituciones vinculadas",
 	"accounts_edit_success": "Cuenta actualizada",
 	"accounts_edit_failed": "No se pudo actualizar la cuenta",
 	"accounts_edit_error_not_found": "Cuenta no encontrada",
@@ -189,6 +179,7 @@
 	"accounts_overview_view_all_positions_other": "Ver las {count} posiciones",
 	"accounts_label_name": "Nombre",
 	"accounts_label_institution": "Institución",
+	"accounts_label_account_number": "Número de cuenta",
 	"accounts_label_category": "Categoría",
 	"accounts_label_balance_group": "Grupo de balance",
 	"accounts_label_currency": "Moneda",
@@ -648,12 +639,13 @@
 	"settings_agent_access_section_title": "Agentes de IA",
 	"settings_agent_access_url_label": "URL de instrucciones",
 	"settings_connections_section_title": "Conexiones",
+	"settings_connections_nav_title": "Instituciones vinculadas",
 	"settings_connections_empty": "Todavía no hay bancos conectados",
 	"settings_connections_unnamed_institution": "Institución sin nombre",
-	"settings_connections_schedule": "Se sincroniza a diario a las {time}",
 	"settings_connections_table_header_institution": "Institución",
 	"settings_connections_table_header_status": "Estado",
 	"settings_connections_table_header_last_synced": "Última sincronización",
+	"settings_connections_table_header_next_sync": "Próxima sincronización automática",
 	"settings_connections_table_header_accounts": "Cuentas",
 	"settings_connections_status_ok": "Conectada",
 	"settings_connections_status_error": "La última sincronización tuvo errores",
