+++ pocketbase/pb_migrations/1786285425_updated_accounts.js
@@ -1,0 +1,30 @@
+/// <reference path="../pb_data/types.d.ts" />
+migrate((app) => {
+  const collection = app.findCollectionByNameOrId("pbc_2324088501")
+
+  // add field
+  collection.fields.addAt(15, new Field({
+    "autogeneratePattern": "",
+    "help": "",
+    "hidden": false,
+    "id": "text840751932",
+    "max": 0,
+    "min": 0,
+    "name": "externalMask",
+    "pattern": "",
+    "presentable": false,
+    "primaryKey": false,
+    "required": false,
+    "system": false,
+    "type": "text"
+  }))
+
+  return app.save(collection)
+}, (app) => {
+  const collection = app.findCollectionByNameOrId("pbc_2324088501")
+
+  // remove field
+  collection.fields.removeById("text840751932")
+
+  return app.save(collection)
+})
